Issue #5337: Rename AutomaticBean
