+++ config/pmd-main.xml
@@ -43,4 +43,11 @@
                         //MethodDeclaration[@Name='iterateAxis']"/>
     </properties>
   </rule>
+  <rule ref="category/java/design.xml/AbstractClassWithoutAnyMethod">
+    <properties>
+    <!-- extends AbstractAutomaticBean, will be removed in later version -->
+      <property name="violationSuppressXPath"
+         value="//ClassOrInterfaceDeclaration[@SimpleName='AutomaticBean']"/>
+    </properties>
+  </rule>
 </ruleset>
