+++ src/main/java/com/puppycrawl/tools/checkstyle/DefaultLogger.java
@@ -39,7 +39,7 @@ import com.puppycrawl.tools.checkstyle.api.SeverityLevel;
  *
  * @see XMLLogger
  */
-public class DefaultLogger extends AutomaticBean implements AuditListener {
+public class DefaultLogger extends AbstractAutomaticBean implements AuditListener {
 
     /**
      * A key pointing to the add exception
@@ -75,10 +75,26 @@ public class DefaultLogger extends AutomaticBean implements AuditListener {
      *
      * @param outputStream where to log audit events
      * @param outputStreamOptions if {@code CLOSE} that should be closed in auditFinished()
+     * @noinspection deprecation
+     * @noinspectionreason We are forced to keep AutomaticBean compatability
+     *     because of maven-checkstyle-plugin.
      */
-    public DefaultLogger(OutputStream outputStream, OutputStreamOptions outputStreamOptions) {
+    public DefaultLogger(OutputStream outputStream,
+                         AutomaticBean.OutputStreamOptions outputStreamOptions) {
+        this(outputStream, OutputStreamOptions.valueOf(outputStreamOptions.name()));
+    }
+
+    /**
+     * Creates a new {@code DefaultLogger} instance.
+     *
+     * @param outputStream where to log audit events
+     * @param outputStreamOptions if {@code CLOSE} that should be closed in auditFinished()
+     */
+    public DefaultLogger(OutputStream outputStream,
+                         AbstractAutomaticBean.OutputStreamOptions outputStreamOptions) {
         // no need to close oS twice
-        this(outputStream, outputStreamOptions, outputStream, OutputStreamOptions.NONE);
+        this(outputStream, outputStreamOptions, outputStream,
+                OutputStreamOptions.NONE);
     }
 
     /**
