+++ src/main/java/com/puppycrawl/tools/checkstyle/XpathFileGeneratorAstFilter.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 
 import com.puppycrawl.tools.checkstyle.api.AuditEvent;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.Violation;
 import com.puppycrawl.tools.checkstyle.xpath.XpathQueryGenerator;
 
@@ -34,7 +33,7 @@ import com.puppycrawl.tools.checkstyle.xpath.XpathQueryGenerator;
  * for {@code XpathFileGeneratorAuditListener}.
  * See issue <a href="https://github.com/checkstyle/checkstyle/issues/102">#102</a>
  */
-public class XpathFileGeneratorAstFilter extends AutomaticBean implements TreeWalkerFilter {
+public class XpathFileGeneratorAstFilter extends AbstractAutomaticBean implements TreeWalkerFilter {
 
     /** The delimiter between xpath queries. */
     private static final String DELIMITER = " | \n";
