+++ src/main/java/com/puppycrawl/tools/checkstyle/api/AbstractViolationReporter.java
@@ -21,6 +21,8 @@ package com.puppycrawl.tools.checkstyle.api;
 
 import java.util.Map;
 
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean;
+
 /**
  * Serves as an abstract base class for all modules that report inspection
  * findings. Such modules have a Severity level which is used for the
@@ -32,7 +34,7 @@ import java.util.Map;
  *      by demand in subclasses
  */
 public abstract class AbstractViolationReporter
-    extends AutomaticBean {
+    extends AbstractAutomaticBean {
 
     /** The severity level of any violations found. */
     private SeverityLevel severityLevel = SeverityLevel.ERROR;
