+++ src/main/java/com/puppycrawl/tools/checkstyle/filters/SuppressionXpathSingleFilter.java
@@ -21,9 +21,9 @@ package com.puppycrawl.tools.checkstyle.filters;
 
 import java.util.regex.Pattern;
 
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean;
 import com.puppycrawl.tools.checkstyle.TreeWalkerAuditEvent;
 import com.puppycrawl.tools.checkstyle.TreeWalkerFilter;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 
 /**
  * <p>
@@ -409,7 +409,7 @@ import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
  *
  * @since 8.18
  */
-public class SuppressionXpathSingleFilter extends AutomaticBean implements
+public class SuppressionXpathSingleFilter extends AbstractAutomaticBean implements
         TreeWalkerFilter {
     /**
      * XpathFilterElement instance.
