+++ src/test/java/com/puppycrawl/tools/checkstyle/MainTest.java
@@ -58,8 +58,8 @@ import org.junit.jupiter.api.io.TempDir;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean.OutputStreamOptions;
 import com.puppycrawl.tools.checkstyle.api.AuditListener;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
 import com.puppycrawl.tools.checkstyle.api.Violation;
 import com.puppycrawl.tools.checkstyle.internal.testmodules.TestRootModuleChecker;
@@ -459,11 +459,11 @@ public class MainTest {
     @Test
     public void testGetOutputStreamOptionsMethod() throws Exception {
         final Path path = new File(getPath("InputMain.java")).toPath();
-        final AutomaticBean.OutputStreamOptions option =
+        final OutputStreamOptions option =
                 TestUtil.invokeStaticMethod(Main.class, "getOutputStreamOptions", path);
         assertWithMessage("Main.getOutputStreamOptions return CLOSE on not null Path")
                 .that(option)
-                .isEqualTo(AutomaticBean.OutputStreamOptions.CLOSE);
+                .isEqualTo(OutputStreamOptions.CLOSE);
     }
 
     @Test
@@ -1826,7 +1826,7 @@ public class MainTest {
     public void testXmlOutputFormatCreateListener() throws IOException {
         final ByteArrayOutputStream out = new ByteArrayOutputStream();
         final AuditListener listener = Main.OutputFormat.XML.createListener(out,
-                AutomaticBean.OutputStreamOptions.CLOSE);
+                OutputStreamOptions.CLOSE);
         assertWithMessage("listener is XMLLogger")
                 .that(listener)
                 .isInstanceOf(XMLLogger.class);
@@ -1836,7 +1836,7 @@ public class MainTest {
     public void testSarifOutputFormatCreateListener() throws IOException {
         final ByteArrayOutputStream out = new ByteArrayOutputStream();
         final AuditListener listener = Main.OutputFormat.SARIF.createListener(out,
-                AutomaticBean.OutputStreamOptions.CLOSE);
+                OutputStreamOptions.CLOSE);
         assertWithMessage("listener is SarifLogger")
                 .that(listener)
                 .isInstanceOf(SarifLogger.class);
@@ -1846,7 +1846,7 @@ public class MainTest {
     public void testPlainOutputFormatCreateListener() throws IOException {
         final ByteArrayOutputStream out = new ByteArrayOutputStream();
         final AuditListener listener = Main.OutputFormat.PLAIN.createListener(out,
-                AutomaticBean.OutputStreamOptions.CLOSE);
+                OutputStreamOptions.CLOSE);
         assertWithMessage("listener is DefaultLogger")
                 .that(listener)
                 .isInstanceOf(DefaultLogger.class);
