+++ src/test/java/com/puppycrawl/tools/checkstyle/XpathFileGeneratorAuditListenerTest.java
@@ -28,8 +28,8 @@ import java.nio.charset.StandardCharsets;
 
 import org.junit.jupiter.api.Test;
 
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean.OutputStreamOptions;
 import com.puppycrawl.tools.checkstyle.api.AuditEvent;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.FileContents;
 import com.puppycrawl.tools.checkstyle.api.FileText;
@@ -96,7 +96,8 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testFinishLocalSetup() {
         final OutputStream out = new ByteArrayOutputStream();
         final XpathFileGeneratorAuditListener listener =
-                new XpathFileGeneratorAuditListener(out, AutomaticBean.OutputStreamOptions.CLOSE);
+                new XpathFileGeneratorAuditListener(out,
+                        OutputStreamOptions.CLOSE);
 
         listener.finishLocalSetup();
         listener.auditStarted(null);
@@ -111,7 +112,8 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testFileStarted() {
         final OutputStream out = new ByteArrayOutputStream();
         final XpathFileGeneratorAuditListener listener =
-                new XpathFileGeneratorAuditListener(out, AutomaticBean.OutputStreamOptions.CLOSE);
+                new XpathFileGeneratorAuditListener(out,
+                        OutputStreamOptions.CLOSE);
         final AuditEvent ev = new AuditEvent(this, "Test.java", null);
         listener.fileStarted(ev);
         listener.auditFinished(null);
@@ -125,7 +127,8 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testFileFinished() {
         final OutputStream out = new ByteArrayOutputStream();
         final XpathFileGeneratorAuditListener listener =
-                new XpathFileGeneratorAuditListener(out, AutomaticBean.OutputStreamOptions.CLOSE);
+                new XpathFileGeneratorAuditListener(out,
+                        OutputStreamOptions.CLOSE);
         final AuditEvent ev = new AuditEvent(this, "Test.java", null);
         listener.fileFinished(ev);
         listener.auditFinished(null);
@@ -139,7 +142,8 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testAddException() {
         final OutputStream out = new ByteArrayOutputStream();
         final XpathFileGeneratorAuditListener logger =
-                new XpathFileGeneratorAuditListener(out, AutomaticBean.OutputStreamOptions.CLOSE);
+                new XpathFileGeneratorAuditListener(out,
+                        OutputStreamOptions.CLOSE);
         logger.auditStarted(null);
         final Violation violation =
                 new Violation(1, 1,
@@ -247,7 +251,7 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testCloseStream() {
         final XpathFileGeneratorAuditListener listener =
                 new XpathFileGeneratorAuditListener(outStream,
-                        AutomaticBean.OutputStreamOptions.CLOSE);
+                        OutputStreamOptions.CLOSE);
         listener.finishLocalSetup();
         listener.auditStarted(null);
         listener.auditFinished(null);
@@ -261,7 +265,7 @@ public class XpathFileGeneratorAuditListenerTest {
     public void testNoCloseStream() {
         final XpathFileGeneratorAuditListener listener =
                 new XpathFileGeneratorAuditListener(outStream,
-                        AutomaticBean.OutputStreamOptions.NONE);
+                        OutputStreamOptions.NONE);
         listener.finishLocalSetup();
         listener.auditStarted(null);
         listener.auditFinished(null);
@@ -318,7 +322,8 @@ public class XpathFileGeneratorAuditListenerTest {
         final TestByteArrayOutputStream out = new TestByteArrayOutputStream();
 
         final XpathFileGeneratorAuditListener listener =
-                new XpathFileGeneratorAuditListener(out, AutomaticBean.OutputStreamOptions.CLOSE);
+                new XpathFileGeneratorAuditListener(out,
+                        OutputStreamOptions.CLOSE);
 
         for (AuditEvent event : events) {
             listener.addError(event);
