+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/utils/TestUtil.java
@@ -34,12 +34,12 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean;
 import com.puppycrawl.tools.checkstyle.PackageNamesLoader;
 import com.puppycrawl.tools.checkstyle.PackageObjectFactory;
 import com.puppycrawl.tools.checkstyle.TreeWalkerAuditEvent;
 import com.puppycrawl.tools.checkstyle.TreeWalkerFilter;
 import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 
@@ -142,7 +142,7 @@ public final class TestUtil {
     }
 
     /**
-     * Checks if stateful field is cleared during {@link AutomaticBean}'s finishLocalSetup.
+     * Checks if stateful field is cleared during {@link AbstractAutomaticBean}'s finishLocalSetup.
      *
      * @param filter filter object which field is to be verified
      * @param event event to pass into filter methods
