+++ src/test/resources/com/puppycrawl/tools/checkstyle/meta/javadocmetadatascraper/InputJavadocMetadataScraperSuppressWarningsFilter.java
@@ -7,10 +7,8 @@ writeXmlOutput = false
 
 package com.puppycrawl.tools.checkstyle.meta.javadocmetadatascraper;
 
-import com.puppycrawl.tools.checkstyle.api.AuditEvent;
-import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
+import com.puppycrawl.tools.checkstyle.AbstractAutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.Filter;
-import com.puppycrawl.tools.checkstyle.checks.SuppressWarningsHolder;
 
 /**
  * <p>
@@ -102,7 +100,7 @@ import com.puppycrawl.tools.checkstyle.checks.SuppressWarningsHolder;
  * @since 5.7
  */
 public abstract class InputJavadocMetadataScraperSuppressWarningsFilter
-    extends AutomaticBean
+    extends AbstractAutomaticBean
     implements Filter {
 
 }
